requirements.txt
# Telegram Bot
aiogram>=3.13.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.36
alembic>=1.13.3
asyncpg>=0.29.0
pgvector>=0.3.5

# Auth
python-jose[cryptography]>=3.3.0

# Scheduler
apscheduler>=3.10.4

# HTTP / Parser
httpx>=0.27.2
beautifulsoup4>=4.12.3
playwright>=1.47.0

# Admin Panel
sqladmin>=0.19.0

# Calendar export
icalendar>=6.0.0

# Config
python-dotenv>=1.0.1
pydantic>=2.9.2
pydantic-settings>=2.5.2

# Utilities
aiofiles>=24.1.0
